+++ src/lib/clients/leveldb.ts
@@ -1,16 +1,15 @@
-import level from 'level';
-import type { LevelUp } from 'levelup';
+import { Level } from 'level';
 import { Config } from './config.js';
 
 export class LevelDB {
-  public static db: LevelUp; // Only public for testing. Not to be used directly outside of this class
+  public static db: Level; // Only public for testing. Not to be used directly outside of this class
 
   // Must be called before other methods (on startup)
-  public static async initialize(imp: any /* for testing */ = level) {
-    LevelDB.db = imp(Config.getConfig().state_db || 'state.ldb', { valueEncoding: 'json' });
+  public static async initialize(imp: any /* for testing */ = Level) {
+    LevelDB.db = new imp(Config.getConfig().state_db || 'state.ldb', { valueEncoding: 'json' });
   }
 
-  public static get(key: string) {
+  public static get(key: string): any {
     return LevelDB.db.get(key);
   }
 
@@ -22,14 +21,8 @@ export class LevelDB {
     return LevelDB.db.del(key);
   }
 
-  public static async list() {
-    return new Promise<any[]>((resolve, reject) => {
-      const stream = LevelDB.db.createValueStream();
-      const data: any[] = [];
-      stream.on('error', reject);
-      stream.on('data', (val) => data.push(val));
-      stream.on('end', () => resolve(data));
-    });
+  public static async list(): Promise<any[]> {
+    return LevelDB.db.values().all();
   }
 
   public static async shutDown() {