+++ CHANGELOG.md
@@ -3,8 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## v5.1.0
+### Added
+- Retry AB torrent upload if received error during previous attempt
+
+### Changed
+- Use more friendly error messages in torrent fetcher
+
 ## v5.0.1
-## Fixed
+### Fixed
 - Fix non-standard "x" chars in valid resolutions
 
 ## v5.0.0